+++ ChangeLog.md
@@ -3,6 +3,7 @@
 ## Version [3.0.0](https://github.com/cedx/sql.net/compare/v2.0.0...v3.0.0)
 - Breaking change: removed some parameters from the `Parameter` primary constructor.
 - Breaking change: renamed the `New-Mapper` cmdlet to `Get-Mapper`.
+- Breaking change: renamed the `-Transaction` parameter of `Close-Connection`, `Approve-Transaction` and `Deny-Transaction` cmdlets to `-InputObject`.
 - Breaking change: transformed the `Mapper` class into a singleton instance.
 - Added new implicit conversions to the `Parameter` class.
 - Added a new overload to the `ParameterCollection` constructor.